Advice for adding properties to messages on topics
When trying to apply this example to messages being published to topics, we found that messages were never received by subscribers. This was tracked down to the default value for pmo.PubLevel being 0, so it's impossible for a subscriber's SubLevel to be less than or equal to this value.
Alternatively - the default value on a new PMO object PubLevel could be updated to match the MQ default value "9".